requirements.txt
# Core dependencies for model connection
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Optional ML frameworks (uncomment as needed)
# torch>=2.0.0
# tensorflow>=2.13.0
# transformers>=4.30.0
# openai>=1.0.0
# anthropic>=0.7.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0